+++ packages/lib-classifier/src/components/Classifier/hooks/index.js
@@ -1,0 +1,2 @@
+export { default as useStore } from './useStore'
+export { default as useWorkflowSnapshot } from './useWorkflowSnapshot'
